Move type definitions and delete outdated comments
