+++ frontend/packages/portal-frontend/src/correlationAnalysis/utilities/helper.ts
@@ -1,13 +1,4 @@
-export interface SortedCorrelations {
-  id: string;
-  feature: string;
-  dose: string | undefined;
-  featureDataset: string;
-  correlation: number;
-  log10qvalue: number;
-  rank: number;
-  [key: string]: any; // in case other keys added
-}
+import { SortedCorrelations } from "../models/CorrelationPlot";
 
 export function transformAndGroupByDataset(
   associationsData: {
